LayoutTests/pointerevents/ios/pointerdown-without-prevent-default-allows-page-scrolling.html

<!DOCTYPE html>
<html>
<head>
<meta charset=utf-8>
<meta name="viewport" content="width=device-width, initial-scale=1">
</head>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../utils.js"></script>
<script>

'use strict';

target_test({ width: "200px", height: "200px" }, (target, test) => {
    document.body.style.width = "2000px";
    document.body.style.height = "2000px";

    document.addEventListener("touchstart", () => {}, { passive: true });
    target.addEventListener("pointerdown", () => {});

    ui.swipe({ x: 100, y: 150 }, { x: 100, y: 50 }).then(() => {
        assert_not_equals(window.pageYOffset, 0, "The page was scrolled in the y-axis.");
        test.done();
    });
}, "Page scrolling is not blocked when pointerdown does not call preventDefault() with passive touch listeners.");

</script>
</body>
</html>
